Code Review

+++ routes/order_products.js
@@ -1,9 +1,10 @@
 const express = require('express');
 const orderProductsRouter = express.Router();
 
-const {getOrderById, getOrderProductById, updateOrderProduct, destroyOrderProduct} = require('../db');
+const {getOrderById, getOrderProductById, updateOrderProduct, destroyOrderProduct, getAllOrderProducts} = require('../db');
+const {requireUser} = require('./utils');
 
-orderProductsRouter.patch('/:orderProductId', async (req, res, next) => {
+orderProductsRouter.patch('/:orderProductId', requireUser, async (req, res, next) => {
     const {price, quantity} = req.body;
     const {orderProductId} = req.params;
     const updateFields = {};
@@ -22,14 +23,26 @@ orderProductsRouter.patch('/:orderProductId', async (req, res, next) => {
 
         if (order.userId === req.user.id) {
             const orderProduct = await updateOrderProduct({id: orderProductId, ...updateFields});
-        
+            
             res.send(orderProduct);
         } else {
-            next({message: 'You are not authorized to access this route.'})
-        }      
+            res.status(401).send({message: 'You are not authorized to access this route.'});
+        }
     } catch (error) {
         next(error);
     }
+})
+
+orderProductsRouter.get('/:orderId', async (req,res,next) => {
+  // take orderId parameter, get all order_products for order (getAllOrderProducts)
+
+  try {
+    const {orderId} = req.params;
+    const order_products = await getAllOrderProducts(orderId);
+    res.send(order_products);
+  } catch (error) {
+    next(error);
+  }
 })
 
 orderProductsRouter.delete('/:orderProductId', async (req, res, next) => {
@@ -39,12 +52,12 @@ orderProductsRouter.delete('/:orderProductId', async (req, res, next) => {
         const product = await getOrderProductById(Number(orderProductId));
         const order = await getOrderById(product.orderId);
 
-        if (order.userId === req.user.id) {
+        if (order && order.userId === req.user.id) {
             const orderProduct = await destroyOrderProduct(orderProductId);
 
             res.send(orderProduct);
         } else {
-            next({message: 'You are not authorized to access this route.'})
+            res.status(401).send({message: 'You are not authorized to access this route.'});
         }
     } catch (error) {
         next(error);
